+++ apps/bbl-nextjs/app/performance/handling-script/afterInteractive.tsx
