+++ apps/bbl-nextjs/app/performance/handling-script/beforeInteractive.tsx
