+++ apps/bbl-nextjs/app/performance/handling-script/lazyOnload.tsx
