+++ apps/bbl-nextjs/app/performance/handling-script/worker.tsx
