--- apps/bbl-nextjs/src/_pages/about/index.tsx
@@ -1,12 +1,0 @@
-import Layout from '../../components/layout';
-import { AboutTemplate } from '@bbl-nx/features';
-
-const AboutPage = () => {
-  return (
-    <Layout>
-      <AboutTemplate />
-    </Layout>
-  );
-};
-
-export default AboutPage;
