--- apps/bbl-nextjs/src/_pages/api/preview.ts
@@ -1,6 +1,0 @@
-import { NextApiRequest, NextApiResponse } from 'next';
-
-export default function handler(__: NextApiRequest, res: NextApiResponse) {
-  res.setPreviewData({});
-  res.end('Preview mode enabled');
-}
