--- apps/bbl-nextjs/src/_pages/index.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { HomeTemplate } from '@bbl-nx/ui-components';
-import Layout from '../components/layout';
-
-const HomePage = () => {
-  return (
-    <Layout>
-      <HomeTemplate />
-    </Layout>
-  );
-};
-
-export default HomePage;
