--- apps/bbl-nextjs/src/_pages/project/index.tsx
@@ -1,12 +1,0 @@
-import { ProjectTemplate } from '@bbl-nx/features';
-import Layout from '../../components/layout';
-
-const ProjectPage = () => {
-  return (
-    <Layout>
-      <ProjectTemplate />
-    </Layout>
-  );
-};
-
-export default ProjectPage;
