+++ libs/ui-components/project.json
@@ -12,7 +12,7 @@
         "outputPath": "dist/libs/ui-components",
         "tsConfig": "libs/ui-components/tsconfig.lib.json",
         "project": "libs/ui-components/package.json",
-        "entryFile": "libs/ui-components/src/index.ts",
+        "entryFile": "libs/ui-components/src/lib/index.ts",
         "external": ["react/jsx-runtime"],
         "rollupConfig": "@nrwl/react/plugins/bundle-rollup",
         "compiler": "babel",
